Refactor navigation links in klettergerüst.php and devteam.php
This commit refactors the navigation links in the klettergerüst.php and devteam.php files. It adds a conditional statement to display a badge with the cart count only if the user is logged in. Additionally, it fixes the formatting of the SVG icons in the devteam.php and kontakt.php links.